test input validation
